feat(transaction): add useTransactionAccordion hook for accordion state management
