+++ src/modules/transactions/hooks/create/index.ts
@@ -1,2 +1,3 @@
 export * from './useCreateTransaction';
 export * from './useCreateTransactionForm';
+export * from './useTransactionAccordion';
